Fix race condition when redeeming an ecash token
There was a TOC-TOU issue with the deduplication of ecash tokens,
because we whould check first, then redeem (which takes a long time
because the library needs to connect to the mint) and then mark as
processed.

This meant that we could still accidentally process the same token
multiple times, causing the "blinded message already signed".

This fixes the issue by atomically checking/marking the token as
processed. Plus it simplifies the migration logic which was incredibly
complex for no reason. We have a way to do migrations, we should just
use it rather than running migrations randomly inside of other actions.